Removed Note Displaying Safari Incompatibility (#6)
* Update README.md

* Update script.js

* Update README.md

* Update README.md

* Added auto-answer functionality

I left the original functionality intact but added the option to use ALT + S to actually answer the questions in a fully automated fashion. Code could be compressed in future if depreciated functionality was removed.

* Made alert prompts more clear

* Updates README for new features

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Fixed undefined for some questions

Certain questions were returning undefined. Turns out that was just because the mergeLists function didn't equate for commas too

* Delete script.js

* Update README.md

* Delete README.md

* Delete console.js

* Revert "Delete console.js"

This reverts commit 97ac26f53ed439e25cebc4a046cd76983793e2da.

* Added Learning Functionality

* Delete .gitignore

* Update .gitignore

* Update .gitignore

* made the semi-automatic mode more semi-automatic

* checked off learning from mistakes

* Update console.js

* combined tampermonkey codebase and console codebase, and updated readme

* fixed error with learning functionality

* fixes bug where it would not learn the answer if the question had semi-colons in it

* Update .gitignore

* Delete .DS_Store

* updated version number

* added a delay to the reading of the answer

EP has the answer as "bleu" for a split second for every question before it actually fetches the right answer, so this delay just gives it a bit of time to do that

* switch from explanation-button to submit-button

for some reason the button being pressed on submit wasn't the submit button, it was the button asking for help. dunno how that got through but it's fixed now ig

* tryna fix an error

* better debugging with catch function

* idk why it only works with the tag name but it only works with the tag name so we'll be using the tag name

* added a split for |

* fixing timings for auto-answer

* Delete script.js

* Update README.md