Added Safari Support
Safari requires the clipboard api to be run in a secure context, and that's not an issue for this script, so it was added. I also discovered a bug where if you start the semi-auto, turn it off, start the fully-auto, turn it off and then go back to semi-auto, it bugs out. Dunno why, but it exists and i can't seem to find a solution (it was there before these code changes)